생성자로 각각의 객체를 필요로 하던 걸 넘겨주는 명시적 DI(의존성 주입 구조에서) Service Locator에 저장하고 필요한 애들만 받아서 하는 구조로 변경

+++ MRenderer/Renderer.cpp
@@ -1459,7 +1459,7 @@ void Renderer::CreateQuadVB()
 
 void Renderer::CreateQuadIB()
 {
-	m_QuadIndexCounts = quadIndices.size();
+	m_QuadIndexCounts = static_cast<UINT>(quadIndices.size());
 	CreateIndexBuffer(m_pDevice.Get(), quadIndices.data(), static_cast<UINT>(quadIndices.size() * sizeof(UINT)), m_QuadIndexBuffers.GetAddressOf());
 }
